+++ app/initializers/ember-pretenderify.js
@@ -1,6 +1,7 @@
 import ENV from '../config/environment';
 import userConfig from '../pretender/config';
 import Server from 'ember-pretenderify/server';
+import readData from 'ember-pretenderify/utils/read-data';
 
 export default {
   name: 'ember-pretenderify',
@@ -10,13 +11,19 @@ export default {
     var shouldUsePretender = config.force || !config.usingProxy;
     var usingInDev = env === 'development' && shouldUsePretender;
     var usingInTest = env === 'test';
+    var shouldUseServer = usingInDev || usingInTest;
 
-    if (usingInDev || usingInTest) {
-      new Server({
-        environment: env,
-        modulePrefix: ENV.modulePrefix,
-        userConfig: userConfig
+    if (shouldUseServer) {
+      var server = new Server({
+        environment: env
       });
+
+      server.loadConfig(userConfig);
+
+      if (usingInDev) {
+        var userData = readData(ENV.modulePrefix);
+        server.loadData(userData);
+      }
     }
   }
 };